Split remote tmux sizing files for Swift budget
