+++ Sources/RemoteTmuxControlConnection.swift
@@ -32,7 +32,7 @@ final class RemoteTmuxControlConnection {
     /// Multicast observer registry. A single connection is shared by every consumer
     /// of the same host+session (``RemoteTmuxController.attach`` reuses it), so events
     /// fan out to all consumers via this registry.
-    private let observers = RemoteTmuxConnectionObservers()
+    let observers = RemoteTmuxConnectionObservers()
 
     // MARK: Observed state
 
@@ -60,24 +60,24 @@ final class RemoteTmuxControlConnection {
     /// guard on `!exited` keep treating a reconnecting connection as alive.
     var exited: Bool { connectionState == .ended }
     private(set) var sessionId: Int?
-    private(set) var windowsByID: [Int: RemoteTmuxWindow] = [:]
-    private(set) var windowOrder: [Int] = []
-    private(set) var activePaneByWindow: [Int: Int] = [:]
-    private(set) var paneOutputByteCounts: [Int: Int] = [:]
-    private(set) var totalOutputBytes = 0
+    var windowsByID: [Int: RemoteTmuxWindow] = [:]
+    var windowOrder: [Int] = []
+    var activePaneByWindow: [Int: Int] = [:]
+    var paneOutputByteCounts: [Int: Int] = [:]
+    var totalOutputBytes = 0
     /// Last-known foreground classification per pane, kept current by the same
     /// one-shot query + live subscription that drive reflow classification
     /// (`#{alternate_on}` + `#{pane_current_command}`, see
     /// ``requestPaneReflow(paneId:)``). Read at close time to decide whether
     /// killing a mirrored pane/window needs a confirmation dialog — a mirror
     /// surface has no local child process for ghostty's needs-confirm check.
-    private(set) var paneForegroundStates: [Int: PaneForegroundState] = [:]
+    var paneForegroundStates: [Int: PaneForegroundState] = [:]
     /// In-flight close-time activity queries by token (see
     /// ``queryWindowActivity(windowId:completion:)``). Failed with `nil` when the
     /// control stream becomes unusable, so a pending close decision falls back to
     /// the cached classification instead of hanging until a reconnect that may
     /// never come.
-    private var activityQueryCompletions: [UUID: ([Int: PaneForegroundState]?) -> Void] = [:]
+    var activityQueryCompletions: [UUID: ([Int: PaneForegroundState]?) -> Void] = [:]
 
     private var process: Process?
     private var stdinWriter: RemoteTmuxControlPipeWriter?
@@ -91,7 +91,7 @@ final class RemoteTmuxControlConnection {
     private var stderrTask: Task<Void, Never>?
     private var parser = RemoteTmuxControlStreamParser()
     private var ingestTask: Task<Void, Never>?
-    private var pendingCommands: [CommandKind] = []
+    var pendingCommands: [CommandKind] = []
     private var connectionWaiters: [UUID: (Bool) -> Void] = [:]
     /// `false` until the attach command's own `%begin`/`%end` block — always the
     /// FIRST block on each control stream, preceding every notification — has been
@@ -105,7 +105,7 @@ final class RemoteTmuxControlConnection {
 
     /// Stateless pure decoders for control-mode message payloads (pane-state seed,
     /// window reorder, session-gone classification). Holds no state.
-    private let decoding = RemoteTmuxControlMessageDecoding()
+    let decoding = RemoteTmuxControlMessageDecoding()
     /// Bounded ring of recent event labels surfaced through `remote.tmux.state`.
     private let diagnostics = RemoteTmuxConnectionDiagnostics()
 
@@ -125,7 +125,7 @@ final class RemoteTmuxControlConnection {
     /// Last client size applied via ``setClientSize(columns:rows:)``, re-applied
     /// after a reconnect so the resumed session keeps the mirror's grid instead of
     /// reverting to ssh's default 80×24.
-    private var lastClientSize: (columns: Int, rows: Int)?
+    var lastClientSize: (columns: Int, rows: Int)?
     /// The last size ANY writer requested via ``setClientSize(columns:rows:)`` —
     /// the shared dedup baseline for every sizing writer on this connection. A
     /// writer must never dedup against a private cache of what IT last pushed:
@@ -135,8 +135,8 @@ final class RemoteTmuxControlConnection {
     var lastRequestedClientSize: (columns: Int, rows: Int)? { lastClientSize }
     /// Instant of the most recent sizing write on this connection — kept for
     /// diagnostics (how stale is the last size request).
-    private(set) var lastSizingSendAt: ContinuousClock.Instant?
-    private var pendingPostAttachAction: PostAttachAction?
+    var lastSizingSendAt: ContinuousClock.Instant?
+    var pendingPostAttachAction: PostAttachAction?
 
     /// Trailing-edge debounce for `refresh-client -C`. SwiftUI layout settle makes the
     /// rendered grid oscillate (e.g. cols 154→155→156→161→…, ~15 distinct grids in
@@ -154,14 +154,14 @@ final class RemoteTmuxControlConnection {
     /// (milliseconds locally), which would forward nearly every oscillation
     /// frame and reinstate the SIGWINCH storm — the oscillation has no
     /// terminating event to gate on.
-    private var clientSizeDebounceTask: Task<Void, Never>?
-    private static let clientSizeDebounceMs = 180
+    var clientSizeDebounceTask: Task<Void, Never>?
+    static let clientSizeDebounceMs = 180
 
     /// Armed on every transition to `.connected` (first connect AND reconnect) and
     /// consumed by the first size apply that follows; see
     /// ``scheduleAttachRedrawKickIfNeeded()`` for why attach needs a redraw kick.
-    private var pendingAttachRedrawKick = false
-    private var attachRedrawKickTask: Task<Void, Never>?
+    var pendingAttachRedrawKick = false
+    var attachRedrawKickTask: Task<Void, Never>?
     /// Gap between the kick's shrink push and its restore push. Must exceed tmux's
     /// pane-resize coalescing (~250 ms), otherwise the two pushes collapse into a
     /// net-zero size change and no SIGWINCH is ever delivered.
@@ -176,7 +176,7 @@ final class RemoteTmuxControlConnection {
     /// again — and any per-window confirmation predicate can be satisfied
     /// spuriously by an unrelated window already at the shrunken height.
     /// Full evidence + a by-hand exploration: docs/remote-tmux-sizing-timers.md.
-    private static let attachRedrawKickGapMs = 350
+    static let attachRedrawKickGapMs = 350
 
     /// Base reconnect backoff (seconds); doubled each attempt up to ``reconnectMaxDelaySeconds``.
     private static let reconnectBaseDelaySeconds: Double = 1
@@ -200,7 +200,7 @@ final class RemoteTmuxControlConnection {
     /// Subscription-name prefix for per-pane `pane_current_path` (`refresh-client -B`).
     /// The tmux pane id is appended so an inbound `%subscription-changed` can be
     /// routed back to its pane; defined once so the writer and reader can't drift.
-    private static let cwdSubscriptionPrefix = "cmux_cwd_"
+    static let cwdSubscriptionPrefix = "cmux_cwd_"
 
     /// Subscription-name prefix for per-pane reflow classification
     /// (`refresh-client -B`). The subscribed format is
@@ -208,13 +208,13 @@ final class RemoteTmuxControlConnection {
     /// and on every change, so launching/exiting an app (bash → node when claude
     /// starts) re-classifies the pane live. The tmux pane id is appended for
     /// routing, mirroring ``cwdSubscriptionPrefix``.
-    private static let reflowSubscriptionPrefix = "cmux_reflow_"
+    static let reflowSubscriptionPrefix = "cmux_reflow_"
     /// Per-pane subscription that keeps header labels LIVE, mirroring
     /// ``cwdSubscriptionPrefix``: tmux pushes the newly-expanded
     /// `pane-border-format` whenever its value changes (a program retitling
     /// its pane, the running command changing) — the same moments native
     /// tmux redraws its own header row.
-    private static let headerSubscriptionPrefix = "cmux_hdr_"
+    static let headerSubscriptionPrefix = "cmux_hdr_"
 
     /// `ESC[?1049h` — enter the alternate screen, emitted to a mirror surface when
     /// the remote pane is on the alternate screen (see ``capturePane(paneId:)``).
@@ -467,258 +467,18 @@ final class RemoteTmuxControlConnection {
         sendInternal(command, kind: .other)
     }
 
-    /// Sizes the tmux control client to `columns`×`rows` cells (tmux
-    /// `refresh-client -C`) so the remote windows/panes reflow to the rendered
-    /// cmux grid. Without this a freshly attached session stays at ssh's default
-    /// 80×24 and TUIs (claude, claude agents) render mangled. Always records the grid
-    /// (re-applied by ``reseedAfterReconnect()``); sends the live `refresh-client`
-    /// only while `.connected`. No-ops for a degenerate grid.
-    ///
-    /// This is the single sizing entrypoint every remote-tmux render path routes
-    /// through (the single-pane display surface and the multi-pane window mirror),
-    /// so client sizing stays one shared behavior rather than duplicated sends.
-    func setClientSize(columns: Int, rows: Int) {
-        guard columns > 0, rows > 0 else { return }
-        // Remember the grid so a reconnect can re-apply it (a fresh ssh client
-        // otherwise reverts to 80×24 and mangles TUIs). Only send now when actually
-        // connected — while reconnecting/ended there is no live stdin (the send would
-        // silently drop); `reseedAfterReconnect` re-applies the stored size.
-        lastClientSize = (columns, rows)
-        lastSizingSendAt = .now
-        guard connectionState == .connected else { return }
-        // Coalesce the layout-settle oscillation into a single send: (re)arm a short
-        // trailing timer; only the last size in a burst actually goes out. The fired
-        // timer is also the "settled" edge that consumes the attach redraw kick.
-        clientSizeDebounceTask?.cancel()
-        clientSizeDebounceTask = Task { @MainActor [weak self] in
-            do {
-                try await ContinuousClock().sleep(for: .milliseconds(Self.clientSizeDebounceMs))
-            } catch {
-                return
-            }
-            guard let self, self.connectionState == .connected, let size = self.lastClientSize else { return }
-            self.send("refresh-client -C \(size.columns)x\(size.rows)")
-            // This send already applied the stored grid — the deferred first-connect
-            // apply would only duplicate it (a deferred reconnect re-seed must stay).
-            if self.pendingPostAttachAction == .applyClientSize {
-                self.pendingPostAttachAction = nil
-            }
-            // Do NOT re-capture here. A re-capture would run capture-pane before the
-            // remote app (claude) finishes its post-SIGWINCH redraw, snapshotting the
-            // stale pre-resize frame and clobbering the correct redraw — the exact
-            // narrow/overlap/duplicate mangle. A manual resize is clean precisely
-            // because it issues no capture: it lets refresh-client -C → SIGWINCH →
-            // the app's own redraw stream back and paint. Attach does the same.
-            self.scheduleAttachRedrawKickIfNeeded()
-        }
-    }
-
-    /// PER-WINDOW client sizing (`refresh-client -C '@id:WxH'`): sizes ONE
-    /// window for this control client instead of the whole session — each
-    /// mirror owns its window's size, so two mirrored windows never fight
-    /// over a shared value. Measured semantics (tmux 3.7): the pin applies
-    /// exactly when this is the sole client, is sticky against session-wide
-    /// pushes, caps `resize-window` per dimension, and with a co-attached
-    /// real client the window sizes to the per-axis MINIMUM of all live
-    /// pins and the real client — the pin is a ceiling, and %layout-change
-    /// stays authoritative over what we requested. Pins are released by
-    /// clean detach and by server-side client teardown; only a crash leaving
-    /// zero clients freezes them (a later real client heals lazily).
-    ///
-    /// Dedup is per window against the last size ANY writer requested for
-    /// that window. The table doubles as the reconnect reseed source:
-    /// ``reseedAfterReconnect()`` re-pins every window (a fresh ssh client
-    /// otherwise reverts everything to 80×24).
-    ///
-    /// If the server rejects the `@id:` form (`%error` — pre-3.x tmux), the
-    /// connection flips to the session-wide fallback for its lifetime and
-    /// surfaces the degraded mode in diagnostics; callers keep calling this
-    /// method either way.
-    func setWindowSize(windowId: Int, columns: Int, rows: Int) {
-        guard columns > 0, rows > 0 else { return }
-        // Dedup only while per-window sizing is live: on the session-wide
-        // fallback the server holds ONE size, so a window's own last request
-        // being unchanged does not mean the server still has it (another
-        // window may have re-sized the session since).
-        if supportsPerWindowSize, let last = lastWindowSizes[windowId], last == (columns, rows),
-           connectionState == .connected {
-            return
-        }
-        #if DEBUG
-        cmuxDebugLog("remote.rects.claim @\(windowId) \(columns)x\(rows)")
-        #endif
-        // Record BEFORE the old-server fallback branch: the table is also the
-        // hidden-mirror claim ledger (updateClientSize's write-once gate) and
-        // the per-window dedup baseline — skipping it on the fallback would
-        // let every hidden mirror re-push a session-wide size forever.
-        lastWindowSizes[windowId] = (columns, rows)
-        lastSizeRequestWindowId = windowId
-        guard supportsPerWindowSize else {
-            setClientSize(columns: columns, rows: rows)
-            return
-        }
-        lastSizingSendAt = .now
-        guard connectionState == .connected else { return }
-        windowSizeDebounceTasks[windowId]?.cancel()
-        windowSizeDebounceTasks[windowId] = Task { @MainActor [weak self] in
-            do {
-                try await ContinuousClock().sleep(for: .milliseconds(Self.clientSizeDebounceMs))
-            } catch {
-                return
-            }
-            guard let self, self.connectionState == .connected,
-                  let size = self.lastWindowSizes[windowId] else { return }
-            self.sendPerWindowSize(windowId: windowId, columns: size.0, rows: size.1)
-            self.scheduleAttachRedrawKickIfNeeded()
-        }
-    }
-
     /// The last size any writer requested per window — per-window dedup
     /// baseline and the reconnect re-pin table.
-    private(set) var lastWindowSizes: [Int: (Int, Int)] = [:]
+    var lastWindowSizes: [Int: (Int, Int)] = [:]
     /// The most recent window a size was requested for — the deterministic
     /// choice when the old-server fallback must replay ONE size session-wide
     /// (the latest requester is in practice the visible tab).
-    private var lastSizeRequestWindowId: Int?
-    private var windowSizeDebounceTasks: [Int: Task<Void, Never>] = [:]
+    var lastSizeRequestWindowId: Int?
+    var windowSizeDebounceTasks: [Int: Task<Void, Never>] = [:]
     /// Whether the server accepts `refresh-client -C '@id:WxH'`. Flipped to
     /// false on the first `%error` for that form (older tmux); sizing then
     /// degrades to the session-wide client size.
-    private(set) var supportsPerWindowSize = true
-
-    /// Sends the per-window form, tagging the command so an `%error` reply
-    /// can flip the capability off and replay via the session-wide path.
-    private func sendPerWindowSize(windowId: Int, columns: Int, rows: Int) {
-        _ = sendInternal(
-            "refresh-client -C '@\(windowId):\(columns)x\(rows)'",
-            kind: .perWindowSize(windowId)
-        )
-    }
-
-    /// Marks the per-window sizing form unsupported (an `%error` came back
-    /// for it) and replays the affected window's size session-wide so the
-    /// session doesn't stay unsized on old servers.
-    func notePerWindowSizeRejected() {
-        guard supportsPerWindowSize else { return }
-        supportsPerWindowSize = false
-        record("remote.tmux.perWindowSize unsupported; falling back to session-wide client size")
-        // Replay the most recently requested window's size — deterministic,
-        // and in practice the visible tab's. (`.values.first` on a Dictionary
-        // could hand the session a hidden tab's stale claim.)
-        let replay = lastSizeRequestWindowId.flatMap { lastWindowSizes[$0] } ?? lastWindowSizes.values.first
-        if let replay {
-            setClientSize(columns: replay.0, rows: replay.1)
-        }
-    }
-
-    /// One-shot, attach-only: force a real SIGWINCH when the attach size push was a
-    /// no-op, so a running TUI re-renders the current frame at the current width.
-    ///
-    /// Why this exists: tmux's grid stores an app's output as it was RENDERED — rows
-    /// drawn at an earlier window width, or an inline TUI's streaming churn, stay in
-    /// the visible frame verbatim. tmux has no "redraw" command for a control (-CC)
-    /// client (`%output` is an append-only pty copy; tmux never re-streams grid
-    /// cells), and `capture-pane` re-reads the same stale cells, so the ONLY way to
-    /// get a clean current-width frame is the app's own repaint — which apps do on
-    /// SIGWINCH. A real-terminal attach virtually always delivers that SIGWINCH
-    /// because its size differs from the detached window's size. The mirror's attach
-    /// usually does NOT: the window still has the size cmux itself left behind, so
-    /// `refresh-client -C` matches it exactly, no pane resize happens, and the stale
-    /// frame stays until the user manually resizes. This kick closes that one gap by
-    /// sending the same signal a real attach sends: a genuine size change (rows-1),
-    /// then the true size after tmux's resize-coalescing window has passed.
-    ///
-    /// Ordering is safe vs the seed: the kick is scheduled after the capture-pane
-    /// commands, and the tmux server processes commands FIFO, so the app's redraw
-    /// `%output` always lands after (on top of) the seed paint. Skipped entirely when
-    /// the attach push itself changed the window size (that already SIGWINCHes), and
-    /// invisible for plain-shell panes (nothing re-renders, nothing is streamed).
-    private func scheduleAttachRedrawKickIfNeeded() {
-        guard pendingAttachRedrawKick else { return }
-        // Not ready yet (no grid computed / topology not drained): keep the one-shot
-        // armed for the next size apply instead of consuming it uselessly.
-        guard connectionState == .connected, !windowsByID.isEmpty else { return }
-        // The size the attach applied: the session-wide client size, or — on
-        // the per-window path, which never sets lastClientSize — the pushed
-        // size of a window that already matches it (the no-op-push case this
-        // kick exists for).
-        let sessionSize = lastClientSize
-        let perWindowNoOp: (windowId: Int, columns: Int, rows: Int)? = lastWindowSizes
-            .compactMap { id, size -> (Int, Int, Int)? in
-                guard let window = windowsByID[id],
-                      window.width == size.0, window.height == size.1 else { return nil }
-                return (id, size.0, size.1)
-            }
-            .first
-            .map { (windowId: $0.0, columns: $0.1, rows: $0.2) }
-        guard sessionSize != nil || perWindowNoOp != nil else { return }
-        pendingAttachRedrawKick = false
-        if let size = sessionSize {
-            guard size.rows > 2 else { return }
-            // Only kick when some mirrored window ALREADY has the target size — i.e. the
-            // size apply above cannot produce a SIGWINCH for it. (window-size latest makes
-            // every window track the client, so one client-level kick redraws them all.)
-            let windowAlreadyAtTarget = windowsByID.values.contains {
-                $0.width == size.columns && $0.height == size.rows
-            }
-            guard windowAlreadyAtTarget else {
-                #if DEBUG
-                cmuxDebugLog("remote.size.kick skip=windowSizeDiffers target=\(size.columns)x\(size.rows)")
-                #endif
-                return
-            }
-            #if DEBUG
-            cmuxDebugLog("remote.size.kick shrink to \(size.columns)x\(size.rows - 1)")
-            #endif
-            attachRedrawKickTask?.cancel()
-            attachRedrawKickTask = Task { @MainActor [weak self] in
-                guard let self, self.connectionState == .connected else { return }
-                // Bail if the user resized since the kick was scheduled: that resize is a
-                // real size change, so it already delivered the SIGWINCH this kick exists
-                // to force — and a shrink at the captured (now stale) size would flash
-                // wrong dimensions at the remote apps.
-                guard let current = self.lastClientSize, current == size else { return }
-                self.send("refresh-client -C \(size.columns)x\(size.rows - 1)")
-                do {
-                    try await ContinuousClock().sleep(for: .milliseconds(Self.attachRedrawKickGapMs))
-                } catch {
-                    return
-                }
-                guard self.connectionState == .connected else { return }
-                // Restore the CURRENT size (the user may have resized during the gap).
-                let restore = self.lastClientSize ?? size
-                #if DEBUG
-                cmuxDebugLog("remote.size.kick restore to \(restore.columns)x\(restore.rows)")
-                #endif
-                self.send("refresh-client -C \(restore.columns)x\(restore.rows)")
-            }
-            return
-        }
-        guard let kick = perWindowNoOp, kick.rows > 2 else { return }
-        #if DEBUG
-        cmuxDebugLog("remote.size.kick @\(kick.windowId) shrink to \(kick.columns)x\(kick.rows - 1)")
-        #endif
-        attachRedrawKickTask?.cancel()
-        attachRedrawKickTask = Task { @MainActor [weak self] in
-            guard let self, self.connectionState == .connected else { return }
-            // Bail if a newer push targeted this window meanwhile — it was a real
-            // size change and already delivered the SIGWINCH.
-            guard let current = self.lastWindowSizes[kick.windowId],
-                  current == (kick.columns, kick.rows) else { return }
-            self.sendPerWindowSize(windowId: kick.windowId, columns: kick.columns, rows: kick.rows - 1)
-            do {
-                try await ContinuousClock().sleep(for: .milliseconds(Self.attachRedrawKickGapMs))
-            } catch {
-                return
-            }
-            guard self.connectionState == .connected else { return }
-            let restore = self.lastWindowSizes[kick.windowId] ?? (kick.columns, kick.rows)
-            #if DEBUG
-            cmuxDebugLog("remote.size.kick @\(kick.windowId) restore to \(restore.0)x\(restore.1)")
-            #endif
-            self.sendPerWindowSize(windowId: kick.windowId, columns: restore.0, rows: restore.1)
-        }
-    }
+    var supportsPerWindowSize = true
 
     /// Requests the current window list + layouts (used to (re)build topology).
     ///
@@ -850,186 +610,12 @@ final class RemoteTmuxControlConnection {
         send("refresh-client -B \(Self.headerSubscriptionPrefix)\(paneId)")
     }
 
-    /// One-shot query of a pane's working directory (`pane_current_path`),
-    /// delivered to the cwd observers. Guarantees an initial folder for the
-    /// mirrored tab even on tmux builds without control-mode subscriptions.
-    func requestPanePath(paneId: Int) {
-        sendInternal(
-            "display-message -p -t %\(paneId) -F \"#{pane_current_path}\"",
-            kind: .panePath(paneId)
-        )
-    }
-
-    /// The exact `refresh-client -B` line that subscribes `paneId`'s working
-    /// directory. The `name:target:format` argument MUST stay double-quoted:
-    /// tmux's command parser rejects an unquoted `#{…}` mid-argument with
-    /// `parse error: syntax error` (verified on tmux 3.6a), and because the
-    /// result FIFO drops `%error` blocks the subscription would silently never
-    /// exist — the mirrored tab's folder would just never update.
-    static func panePathSubscriptionCommand(paneId: Int) -> String {
-        "refresh-client -B \"\(cwdSubscriptionPrefix)\(paneId):%\(paneId):#{pane_current_path}\""
-    }
-
-    /// Subscribes to live `pane_current_path` changes for `paneId` via tmux
-    /// control-mode `refresh-client -B`, so a remote `cd` updates the mirrored
-    /// tab's folder without polling. tmux emits the value once on subscribe and
-    /// again on every change as `%subscription-changed cmux_cwd_<paneId> … : <path>`.
-    /// Best-effort: on tmux builds that don't support subscriptions the command is
-    /// a no-op and ``requestPanePath(paneId:)`` still supplies the initial folder.
-    func subscribePanePath(paneId: Int) {
-        send(Self.panePathSubscriptionCommand(paneId: paneId))
-    }
-
-    /// Removes the live `pane_current_path` subscription for `paneId` (issued once
-    /// the pane is gone). tmux also drops a dead pane's subscriptions on its own;
-    /// this keeps the client's subscription set tidy across split/close churn.
-    func unsubscribePanePath(paneId: Int) {
-        send("refresh-client -B \(Self.cwdSubscriptionPrefix)\(paneId)")
-    }
-
-    /// One-shot query of a pane's reflow classification (`#{alternate_on}` +
-    /// `#{pane_current_command}`), delivered to the reflow observers. This is the
-    /// REQUIRED initial classifier — `subscribePaneReflow` only guarantees *live*
-    /// updates, and on tmux builds where the `-B` subscription doesn't deliver this
-    /// combined value the surface would otherwise stay at its safe no-reflow default
-    /// forever (shells never reflow). Mirrors ``requestPanePath(paneId:)`` exactly
-    /// (a `display-message` always works where a subscription might not).
-    func requestPaneReflow(paneId: Int) {
-        sendInternal(
-            "display-message -p -t %\(paneId) -F \""
-                + "#{alternate_on}\(PaneForegroundState.fieldSeparator)#{pane_current_command}\"",
-            kind: .paneReflow(paneId)
-        )
-    }
-
-    /// Classifies a raw `#{alternate_on}|#{pane_current_command}` value (from the
-    /// one-shot query or a live subscription), records it as the pane's foreground
-    /// state (for the close-confirmation check), and emits the no-reflow decision.
-    /// No-reflow when on the alternate screen OR the foreground command isn't a known
-    /// plain shell; defaults to no-reflow on an empty/unparseable value (safe).
-    private func classifyAndEmitReflow(paneId: Int, rawValue: String, source: String) {
-        let state = PaneForegroundState(rawValue: rawValue)
-        paneForegroundStates[paneId] = state
-        let noReflow = state.suppressesReflow
-        #if DEBUG
-        cmuxDebugLog(
-            "remote.reflow.classify pane=\(paneId) src=\(source) raw=\"\(rawValue.trimmingCharacters(in: .whitespacesAndNewlines))\" "
-                + "alt=\(state.alternateOn ? 1 : 0) cmd=\"\(state.command)\" noReflow=\(noReflow ? 1 : 0)"
-        )
-        #endif
-        observers.emitPaneReflow(paneId, noReflow)
-    }
-
-    /// The exact `refresh-client -B` line that subscribes `paneId`'s foreground
-    /// classification. Same quoting requirement as
-    /// ``panePathSubscriptionCommand(paneId:)`` — unquoted, tmux rejects the
-    /// `#{…}` with a (silently dropped) parse error and the live classification
-    /// never arrives, so a pane that starts a command after its seed keeps its
-    /// stale idle-shell state and the close confirmation never fires.
-    static func paneReflowSubscriptionCommand(paneId: Int) -> String {
-        "refresh-client -B \"\(reflowSubscriptionPrefix)\(paneId):%\(paneId):"
-            + "#{alternate_on}\(PaneForegroundState.fieldSeparator)#{pane_current_command}\""
-    }
-
-    /// Subscribes to live reflow-classification changes for `paneId` via tmux
-    /// control-mode `refresh-client -B`. The subscribed value is
-    /// `#{alternate_on}|#{pane_current_command}`; tmux emits it once on subscribe
-    /// and again whenever it changes, so a pane that switches between a plain shell
-    /// and an inline TUI (e.g. bash → node when claude launches) is reclassified
-    /// without polling. The mirror surface uses this to decide whether to reflow its
-    /// primary screen on resize (shells reflow; alt-screen / inline-TUI panes do
-    /// not), and the close confirmation uses it to track the active foreground
-    /// command. Best-effort: on tmux builds without subscriptions this is a no-op and
-    /// the surface keeps its safe no-reflow default. See ``subscriptionChanged``
-    /// handling for the parse, and ``PaneForegroundState/plainShellCommands`` for the policy.
-    func subscribePaneReflow(paneId: Int) {
-        send(Self.paneReflowSubscriptionCommand(paneId: paneId))
-    }
-
-    /// Removes the live reflow-classification subscription for `paneId` (issued once
-    /// the pane is gone), mirroring ``unsubscribePanePath(paneId:)``.
-    func unsubscribePaneReflow(paneId: Int) {
-        send("refresh-client -B \(Self.reflowSubscriptionPrefix)\(paneId)")
-    }
-
     /// Format for close-time activity queries: the pane id (for cache refresh and
     /// multi-pane correlation) plus the same `alternate_on`/`pane_current_command`
     /// pair the reflow subscription streams. Quoted by the command builders — see
     /// ``panePathSubscriptionCommand(paneId:)`` for why the quoting is load-bearing.
-    private static let activityQueryFormat = "#{pane_id}\(PaneForegroundState.fieldSeparator)"
+    static let activityQueryFormat = "#{pane_id}\(PaneForegroundState.fieldSeparator)"
         + "#{alternate_on}\(PaneForegroundState.fieldSeparator)#{pane_current_command}"
-
-    /// The `list-panes` line behind ``queryWindowActivity(windowId:completion:)``.
-    static func windowActivityQueryCommand(windowId: Int) -> String {
-        "list-panes -t @\(windowId) -F \"\(activityQueryFormat)\""
-    }
-
-    /// The `display-message` line behind ``queryPaneActivity(paneId:completion:)``.
-    static func paneActivityQueryCommand(paneId: Int) -> String {
-        "display-message -p -t %\(paneId) -F \"\(activityQueryFormat)\""
-    }
-
-    /// Live, close-time query of every pane's foreground state in `windowId`.
-    /// tmux evaluates `pane_current_command` AT QUERY TIME, so a command started
-    /// the instant before ⌘W is already visible — unlike the `%subscription-changed`
-    /// cache, which tmux only re-checks about once a second. Results also refresh
-    /// ``paneForegroundStates`` so the synchronous consumers (batch close,
-    /// workspace close, quit warning) get the freshness for free. `completion` is
-    /// called exactly once, on the main actor; `nil` means the query could not be
-    /// issued or the stream reset first (caller falls back to the cache).
-    func queryWindowActivity(windowId: Int, completion: @escaping ([Int: PaneForegroundState]?) -> Void) {
-        sendActivityQuery(Self.windowActivityQueryCommand(windowId: windowId), completion: completion)
-    }
-
-    /// Single-pane variant of ``queryWindowActivity(windowId:completion:)``, for
-    /// the multi-pane mirror's pane-header ✕ close.
-    func queryPaneActivity(paneId: Int, completion: @escaping ([Int: PaneForegroundState]?) -> Void) {
-        sendActivityQuery(Self.paneActivityQueryCommand(paneId: paneId), completion: completion)
-    }
-
-    private func sendActivityQuery(
-        _ command: String, completion: @escaping ([Int: PaneForegroundState]?) -> Void
-    ) {
-        guard connectionState == .connected else {
-            completion(nil)
-            return
-        }
-        let token = UUID()
-        activityQueryCompletions[token] = completion
-        guard sendInternal(command, kind: .activityQuery(token)) else {
-            // The stream could not accept the query, so no result can correlate.
-            // Fail now and let the close decision proceed on the cached state.
-            activityQueryCompletions.removeValue(forKey: token)?(nil)
-            return
-        }
-    }
-
-    /// Parses one activity-query line (``activityQueryFormat``):
-    /// `%<paneId>|<alternate_on>|<pane_current_command>`. `nil` for an
-    /// unparseable line — the caller treats that pane as unclassified.
-    /// `maxSplits: 1` is deliberate (NOT 2): this strips only the `%paneId`
-    /// prefix, and ``PaneForegroundState/init(rawValue:)`` applies its own
-    /// `maxSplits: 1` for the second field — so a `|` inside a command name
-    /// stays in the command instead of truncating it.
-    static func parseActivityQueryLine(_ line: String) -> (paneId: Int, state: PaneForegroundState)? {
-        let trimmed = line.trimmingCharacters(in: .whitespacesAndNewlines)
-        let parts = trimmed.split(
-            separator: PaneForegroundState.fieldSeparator, maxSplits: 1, omittingEmptySubsequences: false
-        )
-        guard parts.count == 2,
-              let paneId = RemoteTmuxControlStreamParser.id(parts[0], sigil: "%") else { return nil }
-        return (paneId, PaneForegroundState(rawValue: String(parts[1])))
-    }
-
-    /// Fails every in-flight activity query — called whenever the control stream
-    /// becomes unusable (reconnect begins, deliberate stop, genuine `%exit`), so
-    /// a pending close decision falls back to the cached classification.
-    private func failPendingActivityQueries() {
-        guard !activityQueryCompletions.isEmpty else { return }
-        let completions = Array(activityQueryCompletions.values)
-        activityQueryCompletions.removeAll()
-        for completion in completions { completion(nil) }
-    }
 
     private func finishConnectionWaiters(connected: Bool) {
         guard !connectionWaiters.isEmpty else { return }
@@ -1145,7 +731,7 @@ final class RemoteTmuxControlConnection {
     // MARK: - Internals
 
     @discardableResult
-    private func sendInternal(_ command: String, kind: CommandKind) -> Bool {
+    func sendInternal(_ command: String, kind: CommandKind) -> Bool {
         guard connectionState == .connected, let stdinWriter else { return false }
         let line = command.hasSuffix("\n") ? command : command + "\n"
         guard let data = line.data(using: .utf8) else { return false }
@@ -1282,7 +868,7 @@ final class RemoteTmuxControlConnection {
     /// and re-capture current contents (with history) + cwd. Called from the first
     /// post-reconnect `list-windows` result, so `windowsByID` is freshly repopulated
     /// and the command-result FIFO is aligned (the attach block is already drained).
-    private func reseedAfterReconnect() {
+    func reseedAfterReconnect() {
         if let size = lastClientSize {
             send("refresh-client -C \(size.columns)x\(size.rows)")
         }
@@ -1493,285 +1079,6 @@ final class RemoteTmuxControlConnection {
         return idBearingName ?? documentedName
     }
 
-    private func handleCommandResult(lines: [String], isError: Bool) {
-        // The attach block was already consumed upstream (`attachBlockDrained`);
-        // an empty FIFO here means an unsolicited block — drop it rather than
-        // misalign the positional correlation.
-        guard !pendingCommands.isEmpty else { return }
-        let kind = pendingCommands.removeFirst()
-        guard !isError else {
-            // An errored activity query must still complete (with nil) — a close
-            // decision is waiting on it and falls back to the cached state.
-            if case let .activityQuery(token) = kind,
-               let completion = activityQueryCompletions.removeValue(forKey: token) {
-                completion(nil)
-            }
-            // A rejected per-window size normally means the server predates
-            // the '@id:WxH' form: degrade to session-wide sizing, visibly.
-            // But a "can't find window" error is about ONE dead window (it
-            // raced a close) — drop that entry instead of downgrading the
-            // whole connection.
-            if case let .perWindowSize(windowId) = kind {
-                if lines.joined(separator: " ").localizedCaseInsensitiveContains("find window") {
-                    lastWindowSizes[windowId] = nil
-                } else {
-                    notePerWindowSizeRejected()
-                }
-            }
-            // An errored rects fetch still owes its pending layout a
-            // resolution: retry once, then drop the pending tree so
-            // observers keep the last VERIFIED layout (never a raw one).
-            if case let .paneRects(windowId, generation) = kind {
-                handlePaneRectsFailure(windowId: windowId, generation: generation)
-            }
-            // Errors are dropped by design (results correlate positionally), but
-            // an invisible %error has already hidden one real bug — an unquoted
-            // refresh-client -B that never subscribed — so leave a trace.
-            #if DEBUG
-            cmuxDebugLog(
-                "remote.tmux.commandError kind=\(kind) error=\"\(lines.joined(separator: " / "))\""
-            )
-            #endif
-            return
-        }
-        switch kind {
-        case let .paneRects(windowId, generation):
-            handlePaneRectsReply(windowId: windowId, generation: generation, lines: lines)
-        case .listWindows:
-            var order: [Int] = []
-            var next: [Int: RemoteTmuxWindow] = [:]
-            for line in lines {
-                // "@<id> <layout> <visible-layout> [<flags>] <name with spaces…>"
-                // — id and the layout strings never contain spaces; flags are
-                // bracket-delimited because they can be EMPTY (an empty bare
-                // field would collapse under whitespace splitting).
-                let parts = line.split(separator: " ", maxSplits: 4, omittingEmptySubsequences: false)
-                guard parts.count >= 4,
-                      let id = RemoteTmuxControlStreamParser.id(parts[0], sigil: "@"),
-                      let node = RemoteTmuxRawLayoutParser.parse(String(parts[1]))
-                else { continue }
-                let visibleNode = RemoteTmuxRawLayoutParser.parse(String(parts[2]))
-                let flags = String(parts[3]).trimmingCharacters(in: CharacterSet(charactersIn: "[]"))
-                let name = parts.count >= 5 ? String(parts[4]) : ""
-                next[id] = RemoteTmuxWindow(
-                    id: id,
-                    name: name,
-                    width: node.width,
-                    height: node.height,
-                    layout: node,
-                    visibleLayout: visibleNode,
-                    zoomed: flags.contains("Z") && visibleNode != nil
-                )
-                order.append(id)
-            }
-            // `next` holds RAW string geometry — it is never published as-is.
-            // Each window is staged and re-published by its own rects reply;
-            // until then observers keep that window's previous verified tree
-            // (or, for a brand-new window, don't see it yet).
-            // Ignore an empty/garbled reply on purpose: a live tmux session always
-            // has ≥1 window, so a zero-window parse is a transient or malformed
-            // result, not a real topology. Acting on it would wipe `windowOrder`
-            // and tear down every mirror tab. A genuine "no windows" state means
-            // the session ended — that arrives as the connection's `%exit` /
-            // stream-end (see `handleConnectionExited` → `handleSessionEndedRemotely`),
-            // which is the path that closes the workspace / dedicated window.
-            if !order.isEmpty {
-                // Replace topology instead of merging: a remote close missed while
-                // disconnected leaves no %window-close, so prune stale panes here.
-                let liveIDs = Set(order)
-                // REMOVALS and name updates publish now (they carry no leaf
-                // geometry); every window's GEOMETRY is staged and published
-                // only by its rects reply. Verified entries for surviving
-                // windows stay as-is until then.
-                windowsByID = windowsByID.filter { liveIDs.contains($0.key) }
-                pendingLayouts = pendingLayouts.filter { liveIDs.contains($0.key) }
-                // A population that starts from an empty table (first attach,
-                // reconnect reseed after every window closed) publishes
-                // atomically: hold each window's verified tree in staging
-                // until the LAST rects reply lands. Otherwise tab creation
-                // order — and initial selection — would follow reply arrival
-                // order, not tmux's window order.
-                initialBatchStaged = initialBatchStaged.filter { liveIDs.contains($0.key) }
-                if windowsByID.isEmpty {
-                    initialBatchAwaiting = Set(order).subtracting(initialBatchStaged.keys)
-                    #if DEBUG
-                    cmuxDebugLog("remote.rects.batchArm windows=\(order)")
-                    #endif
-                } else if var awaiting = initialBatchAwaiting {
-                    awaiting.formIntersection(liveIDs)
-                    initialBatchAwaiting = awaiting
-                    flushInitialBatchIfDrained()
-                }
-                for (id, window) in next {
-                    if let existing = windowsByID[id], existing.name != window.name {
-                        windowsByID[id] = RemoteTmuxWindow(
-                            id: id, name: window.name,
-                            width: existing.width, height: existing.height,
-                            layout: existing.layout, visibleLayout: existing.visibleLayout,
-                            zoomed: existing.zoomed
-                        )
-                    }
-                    stagePendingLayout(
-                        windowId: id,
-                        node: window.layout, visibleNode: window.visibleLayout,
-                        zoomed: window.zoomed, name: window.name
-                    )
-                }
-                // Per-window sizing state must not outlive the topology: a
-                // stale pin would be replayed by the reconnect reseed, and a
-                // pending debounce could fire at a dead @id.
-                lastWindowSizes = lastWindowSizes.filter { liveIDs.contains($0.key) }
-                for (id, task) in windowSizeDebounceTasks where !liveIDs.contains(id) {
-                    task.cancel()
-                    windowSizeDebounceTasks[id] = nil
-                }
-                if let last = lastSizeRequestWindowId, !liveIDs.contains(last) {
-                    lastSizeRequestWindowId = nil
-                }
-                activePaneByWindow = activePaneByWindow.filter { liveIDs.contains($0.key) }
-                windowTitleRowsVisible = windowTitleRowsVisible.filter { liveIDs.contains($0.key) }
-                prunePaneState(keeping: Set(next.values.flatMap { $0.paneIDsInOrder }))
-                windowOrder = order
-                // Publish removals/order/names; geometry rides each window's
-                // rects reply.
-                observers.notifyTopologyChanged()
-                // The attach block is drained and the topology is fresh — run the
-                // deferred post-attach work; commands queued here correlate cleanly
-                // (see ``PostAttachAction``).
-                switch pendingPostAttachAction {
-                case .reseed:
-                    reseedAfterReconnect()
-                case .applyClientSize:
-                    // A surface that hasn't computed a grid yet is covered by the
-                    // debounced `setClientSize` instead.
-                    if let size = lastClientSize {
-                        send("refresh-client -C \(size.columns)x\(size.rows)")
-                    }
-                case nil:
-                    break
-                }
-                pendingPostAttachAction = nil
-                // First-connect coverage for the attach redraw kick happens at
-                // the publication point (each window's rects reply): here
-                // `windowsByID` is still empty on a first connect — geometry
-                // publishes only when the rects replies land.
-            }
-        case let .capturePane(paneId):
-            // capture-pane -e -S output is the pane's history + visible rows (with
-            // SGR escapes). Home + clear the VISIBLE SCREEN (ESC[2J — NOT ESC[3J,
-            // which would erase the scrollback we are seeding), then write every
-            // captured row joined by CR LF: rows that overflow the screen scroll up
-            // into the surface's scrollback buffer, which is what makes the mirrored
-            // tab scrollable from the start. The last row (the visible bottom) gets
-            // no trailing newline so the cursor lands at its END, lining up with
-            // tmux's real prompt cursor — otherwise echoed input lands a line below
-            // the prompt. The `.paneState` seed then repositions the cursor within
-            // the visible screen.
-            let painted = "\u{1b}[H\u{1b}[2J" + lines.joined(separator: "\r\n")
-            if let data = painted.data(using: .utf8) {
-                observers.emitPaneOutput(paneId, data)
-            }
-        case let .paneState(paneId):
-            // Restore the pane's terminal state (scroll region + DEC modes + cursor)
-            // onto the mirror surface, applied after the capture paint. The scroll
-            // region (DECSTBM) is the important one: without it an inline TUI's
-            // region-relative redraws land on the wrong rows even at a static size.
-            if let line = lines.first {
-                observers.emitPaneOutput(paneId, decoding.paneStateSeedSequence(from: line))
-            }
-        case let .panePath(paneId):
-            if let path = lines.first?.trimmingCharacters(in: .whitespaces), !path.isEmpty {
-                observers.emitPaneCwd(paneId, path)
-            }
-        case let .paneReflow(paneId):
-            // One-shot reflow classification result (see requestPaneReflow). Empty
-            // lines → classifyAndEmitReflow defaults to no-reflow (safe).
-            classifyAndEmitReflow(paneId: paneId, rawValue: lines.first ?? "", source: "oneshot")
-        case let .activityQuery(token):
-            guard let completion = activityQueryCompletions.removeValue(forKey: token) else { break }
-            var states: [Int: PaneForegroundState] = [:]
-            for line in lines {
-                guard let parsed = Self.parseActivityQueryLine(line) else { continue }
-                states[parsed.paneId] = parsed.state
-            }
-            // The fresh answer flows back into the cache, so the synchronous
-            // consumers (batch close, workspace close, quit warning) benefit too.
-            for (paneId, state) in states { paneForegroundStates[paneId] = state }
-            completion(states)
-        case let .paneAltScreen(paneId):
-            // Match the mirror surface to the remote pane's screen (alt = no reflow on
-            // resize). Emitted before the capture paint that follows in the FIFO, so the
-            // seeded rows land on the right screen. The else branch is load-bearing on a
-            // surface REUSED across reconnect: if it was on the alt screen before and the
-            // remote pane is now on primary, force it back (1049l) so the capture doesn't
-            // paint onto a stale alt screen.
-            if lines.first?.trimmingCharacters(in: .whitespaces) == "1" {
-                observers.emitPaneOutput(paneId, Self.altScreenEnterSequence)
-            } else {
-                observers.emitPaneOutput(paneId, Self.altScreenExitSequence)
-            }
-        case .perWindowSize:
-            // A successful per-window size push replies with an empty block;
-            // the interesting outcome (%error -> capability fallback) is
-            // handled in the error branch above.
-            break
-        case .other:
-            break
-        }
-    }
-
-    private func applyLayout(
-        windowId: Int, layout: String, visibleLayout: String? = nil, zoomed: Bool = false
-    ) {
-        guard let node = RemoteTmuxRawLayoutParser.parse(layout) else { return }
-        // Preserve any name tmux already reported (a %layout-change carries no name).
-        let existingName = windowsByID[windowId]?.name ?? pendingLayouts[windowId]?.name ?? ""
-        let visibleNode = visibleLayout.flatMap { RemoteTmuxRawLayoutParser.parse($0) }
-        stagePendingLayout(
-            windowId: windowId,
-            node: node, visibleNode: visibleNode,
-            zoomed: zoomed && visibleNode != nil,
-            name: existingName
-        )
-    }
-
-    /// Quarantines a parsed layout and drives the rects fetch that will
-    /// publish it. Coalesces: while a fetch is in flight, newer layouts just
-    /// replace the pending tree (bumping the generation) and mark it dirty
-    /// for ONE follow-up fetch.
-    private func stagePendingLayout(
-        windowId: Int,
-        node: RemoteTmuxLayoutNode,
-        visibleNode: RemoteTmuxLayoutNode?,
-        zoomed: Bool,
-        name: String
-    ) {
-        var pending = pendingLayouts[windowId] ?? PendingLayout(
-            node: node, visibleNode: visibleNode, zoomed: zoomed, name: name, generation: 0
-        )
-        pending.node = node
-        pending.visibleNode = visibleNode
-        pending.zoomed = zoomed
-        pending.name = name
-        pending.generation += 1
-        pending.retriesRemaining = 1
-        if pending.inFlight {
-            pending.dirty = true
-            pendingLayouts[windowId] = pending
-            return
-        }
-        pending.inFlight = requestPaneRects(windowId: windowId, generation: pending.generation)
-        // Send failure leaves inFlight false. Backpressure already began
-        // reconnecting; the not-connected/no-writer case is recovered the
-        // same way — the next (re)connect's spawn resets this table and the
-        // attach list-windows restages every window. The raw tree stays
-        // quarantined either way.
-        pendingLayouts[windowId] = pending
-        #if DEBUG
-        cmuxDebugLog("remote.rects.stage @\(windowId) gen=\(pending.generation) sent=\(pending.inFlight ? 1 : 0)")
-        #endif
-    }
-
     /// Per-pane header-strip labels: the pane's EXPANDED `pane-border-format`
     /// (style tokens stripped) — exactly the text a native tmux client draws
     /// in that pane's header, custom formats included. Seeded by the
@@ -1779,13 +1086,13 @@ final class RemoteTmuxControlConnection {
     /// (`cmux_hdr_<pane>`), so a program retitling its pane updates the strip
     /// the moment tmux would redraw its own border. The mirror copies its
     /// windows' subset on reconcile; the view never reads this directly.
-    private(set) var paneHeaderLabels: [Int: String] = [:]
+    var paneHeaderLabels: [Int: String] = [:]
 
     /// Whether each window currently has `pane-border-status top` — i.e.
     /// tmux itself is drawing header rows, which is the ONLY time the strips
     /// show label text (a stock tmux displays no titles anywhere; cmux adds
     /// only the active-pane dot on top of that).
-    private(set) var windowTitleRowsVisible: [Int: Bool] = [:]
+    var windowTitleRowsVisible: [Int: Bool] = [:]
 
     /// Drops tmux `#[...]` style tokens from an expanded format (tmux marks
     /// the active pane by reversing its index; the dot carries that signal
@@ -1801,7 +1108,7 @@ final class RemoteTmuxControlConnection {
     /// publishes the pre-title tree), so raw trees are quarantined here and
     /// enter `windowsByID` only patched with list-panes rects — observers
     /// can never see string geometry, structurally.
-    private struct PendingLayout {
+    struct PendingLayout {
         var node: RemoteTmuxLayoutNode
         var visibleNode: RemoteTmuxLayoutNode?
         var zoomed: Bool
@@ -1817,7 +1124,7 @@ final class RemoteTmuxControlConnection {
         var inFlight = false
         var retriesRemaining = 1
     }
-    private var pendingLayouts: [Int: PendingLayout] = [:]
+    var pendingLayouts: [Int: PendingLayout] = [:]
 
     /// Window ids from a topology population that started with NO published
     /// windows (first attach, reconnect reseed into an empty table), still
@@ -1828,193 +1135,10 @@ final class RemoteTmuxControlConnection {
     /// creation order — and with it which tab ends up selected and which
     /// mirrors take their one-time size claim from a hidden, collapsed
     /// container — would be a race between round trips.
-    private var initialBatchAwaiting: Set<Int>?
-    private var initialBatchStaged: [Int: RemoteTmuxWindow] = [:]
+    var initialBatchAwaiting: Set<Int>?
+    var initialBatchStaged: [Int: RemoteTmuxWindow] = [:]
 
-    /// Marks `windowId` resolved (published into staging, dropped, or closed)
-    /// for the initial atomic batch; flushes the batch when it drains.
-    private func finishInitialBatchMember(_ windowId: Int) {
-        guard var awaiting = initialBatchAwaiting else { return }
-        awaiting.remove(windowId)
-        initialBatchAwaiting = awaiting
-        flushInitialBatchIfDrained()
-    }
-
-    private func flushInitialBatchIfDrained() {
-        guard let awaiting = initialBatchAwaiting, awaiting.isEmpty else { return }
-        for (id, window) in initialBatchStaged { windowsByID[id] = window }
-        initialBatchStaged = [:]
-        initialBatchAwaiting = nil
-        prunePaneState(keeping: Set(windowsByID.values.flatMap { $0.paneIDsInOrder }))
-        record("initial-batch-published")
-        #if DEBUG
-        cmuxDebugLog("remote.rects.batchFlush windows=\(windowsByID.keys.sorted())")
-        #endif
-        observers.notifyTopologyChanged()
-        scheduleAttachRedrawKickIfNeeded()
-    }
-
-    /// THE publication point for layout geometry — the module invariant:
-    /// `windowsByID` (what observers read) only ever holds trees whose leaf
-    /// rects came from list-panes. Quarantined layouts (`pendingLayouts`)
-    /// are published here, generation-guarded, or not at all.
-    private func handlePaneRectsReply(windowId: Int, generation: Int, lines: [String]) {
-        #if DEBUG
-        cmuxDebugLog(
-            "remote.rects.reply @\(windowId) gen=\(generation) pendingGen=\(pendingLayouts[windowId]?.generation ?? -1) "
-                + "lines=\(lines.count) awaiting=\(initialBatchAwaiting.map(String.init(describing:)) ?? "nil")"
-        )
-        #endif
-        guard var pending = pendingLayouts[windowId] else {
-            // Window closed while the fetch was in flight; nothing to publish.
-            return
-        }
-        pending.inFlight = false
-        guard generation == pending.generation else {
-            // Stale reply for an older layout. A newer fetch is owed: send it.
-            pending.inFlight = requestPaneRects(windowId: windowId, generation: pending.generation)
-            pending.dirty = false
-            pendingLayouts[windowId] = pending
-            return
-        }
-        var rects: [Int: (x: Int, y: Int, width: Int, height: Int)] = [:]
-        var labels: [Int: String] = [:]
-        var activePane: Int?
-        var titleRowsVisible = false
-        for line in lines {
-            // "%id left top width height active border-status :format…" —
-            // the expanded pane-border-format is last (it may contain
-            // spaces) behind the ':' sentinel (it may be empty).
-            let parts = line.split(separator: " ", maxSplits: 7, omittingEmptySubsequences: false)
-            guard parts.count >= 8,
-                  let paneId = RemoteTmuxControlStreamParser.id(parts[0], sigil: "%"),
-                  let x = Int(parts[1]), let y = Int(parts[2]),
-                  let width = Int(parts[3]), let height = Int(parts[4]),
-                  width > 0, height > 0
-            else { continue }
-            rects[paneId] = (x: x, y: y, width: width, height: height)
-            if parts[5] == "1" { activePane = paneId }
-            // Labels render only where tmux itself draws headers: `top` rows
-            // are the strips above each pane. (`bottom` rows keep faithful
-            // GEOMETRY via the rects, but carry no label — the strip-segment
-            // match keys on pane TOP edges.)
-            if parts[6] == "top" { titleRowsVisible = true }
-            labels[paneId] = Self.strippingStyleTokens(String(parts[7].dropFirst()))
-        }
-        // The reply must cover EVERY pane of the tree it will publish:
-        // `patchingLeafRects` leaves unknown leaves untouched, so a partial
-        // reply (malformed line, zero-sized mid-resize rect, pane closed
-        // between the layout event and this fetch) would smuggle raw
-        // layout-string geometry into `windowsByID` — the exact thing the
-        // quarantine exists to prevent.
-        let requiredPanes = Set(pending.node.paneIDsInOrder)
-            .union(pending.visibleNode.map { Set($0.paneIDsInOrder) } ?? [])
-        guard !rects.isEmpty, requiredPanes.allSatisfy({ rects[$0] != nil }) else {
-            // Garbled/partial reply. Retry once; then drop the pending layout —
-            // observers keep the last VERIFIED tree rather than ever seeing a
-            // raw one.
-            if pending.retriesRemaining > 0 {
-                pending.retriesRemaining -= 1
-                pending.inFlight = requestPaneRects(windowId: windowId, generation: pending.generation)
-                pendingLayouts[windowId] = pending
-            } else {
-                pendingLayouts[windowId] = nil
-                record("pane-rects-dropped @\(windowId)")
-                finishInitialBatchMember(windowId)
-            }
-            return
-        }
-        for (paneId, label) in labels where paneHeaderLabels[paneId] != label {
-            paneHeaderLabels[paneId] = label
-        }
-        if windowTitleRowsVisible[windowId] != titleRowsVisible {
-            windowTitleRowsVisible[windowId] = titleRowsVisible
-        }
-        // The fetch's #{pane_active} is a fresh server snapshot: adopt it
-        // whenever it differs, not only on first sight — an active-pane
-        // change during a disconnect has no %window-pane-changed to replay,
-        // so this is the path that repairs it. A user switch racing this
-        // reply self-corrects: its own %window-pane-changed follows.
-        if let activePane, activePaneByWindow[windowId] != activePane {
-            activePaneByWindow[windowId] = activePane
-            observers.emitActivePaneChanged(windowId, activePane)
-        }
-        let published = RemoteTmuxWindow(
-            id: windowId,
-            name: pending.name,
-            width: pending.node.width,
-            height: pending.node.height,
-            layout: pending.node.patchingLeafRects(rects),
-            visibleLayout: pending.visibleNode?.patchingLeafRects(rects),
-            zoomed: pending.zoomed
-        )
-        if pending.dirty {
-            // A newer layout superseded this one mid-flight: publish this
-            // verified state now (it is true as of this reply) and fetch the
-            // newer generation once.
-            pending.dirty = false
-            pending.inFlight = requestPaneRects(windowId: windowId, generation: pending.generation)
-            pendingLayouts[windowId] = pending
-        } else {
-            pendingLayouts[windowId] = nil
-        }
-        record("pane-rects @\(windowId)")
-        if initialBatchAwaiting != nil {
-            // First population: hold verified windows in staging and publish
-            // them all at once when the last reply lands, so observers never
-            // see a partial topology and tab creation stays deterministic.
-            initialBatchStaged[windowId] = published
-            finishInitialBatchMember(windowId)
-            return
-        }
-        windowsByID[windowId] = published
-        if !windowOrder.contains(windowId) { windowOrder.append(windowId) }
-        prunePaneState(keeping: Set(windowsByID.values.flatMap { $0.paneIDsInOrder }))
-        observers.notifyTopologyChanged()
-        // First-connect coverage for the attach redraw kick: if the grid was
-        // computed before `.enter`, no post-connect `setClientSize` may ever
-        // fire (layout settled + same-size dedupe upstream), so the
-        // debounced-send consumer never runs. This publication is the earliest
-        // point with populated topology; the geometry it holds predates tmux
-        // processing the post-attach size apply, so the at-target check sees
-        // the true pre-apply size. One-shot guarded — no-op when already
-        // consumed (or when reseedAfterReconnect ran it).
-        scheduleAttachRedrawKickIfNeeded()
-    }
-
-    /// Retry-or-drop for a rects fetch that errored: the pending layout must
-    /// never be published raw, and must not dangle in-flight forever.
-    private func handlePaneRectsFailure(windowId: Int, generation: Int) {
-        #if DEBUG
-        cmuxDebugLog("remote.rects.error @\(windowId) gen=\(generation)")
-        #endif
-        guard var pending = pendingLayouts[windowId] else { return }
-        pending.inFlight = false
-        if pending.generation != generation || pending.dirty {
-            // A newer layout is owed a fetch regardless of this failure.
-            pending.dirty = false
-            pending.inFlight = requestPaneRects(windowId: windowId, generation: pending.generation)
-            pendingLayouts[windowId] = pending
-            return
-        }
-        if pending.retriesRemaining > 0 {
-            pending.retriesRemaining -= 1
-            pending.inFlight = requestPaneRects(windowId: windowId, generation: pending.generation)
-            pendingLayouts[windowId] = pending
-        } else {
-            pendingLayouts[windowId] = nil
-            record("pane-rects-dropped @\(windowId)")
-            finishInitialBatchMember(windowId)
-        }
-    }
-
-    private func prunePaneState(keeping livePanes: Set<Int>) {
-        paneHeaderLabels = paneHeaderLabels.filter { livePanes.contains($0.key) }
-        paneOutputByteCounts = paneOutputByteCounts.filter { livePanes.contains($0.key) }
-        paneForegroundStates = paneForegroundStates.filter { livePanes.contains($0.key) }
-    }
-
-    private func record(_ event: String) {
+    func record(_ event: String) {
         diagnostics.record(event)
     }
