+++ cmuxTests/TerminalAndGhosttyTests.swift
@@ -4687,96 +4687,6 @@ final class GhosttySurfaceOverlayTests: XCTestCase {
     }
 }
 
-
-@MainActor
-final class MainWindowSelfSizingTests: XCTestCase {
-    /// The main window must never resize itself to fit its SwiftUI content.
-    /// NSHostingView watches window layout and calls NSWindow.setFrame when
-    /// the measured content size disagrees with the window
-    /// (updateAnimatedWindowSize) — with content whose measured size tracks
-    /// the container, that path grows the window a step per layout pass,
-    /// without bound. MainWindowHostingView disables it (sizingOptions = []);
-    /// this pins that contract with content whose ideal size is far larger
-    /// than the window.
-    @MainActor
-    func testWindowDoesNotGrowTowardContentIdealSize() {
-        let window = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 500, height: 400),
-            styleMask: [.titled, .closable, .resizable],
-            backing: .buffered,
-            defer: false
-        )
-        defer {
-            NotificationCenter.default.post(name: NSWindow.willCloseNotification, object: window)
-            window.orderOut(nil)
-        }
-        let oversized = Color.clear.frame(
-            minWidth: 0, idealWidth: 4000, maxWidth: .infinity,
-            minHeight: 0, idealHeight: 3000, maxHeight: .infinity
-        )
-        window.contentView = MainWindowHostingView(rootView: AnyView(oversized))
-        window.setFrame(NSRect(x: 0, y: 0, width: 500, height: 400), display: true)
-        window.makeKeyAndOrderFront(nil)
-
-        // Several display cycles: the hosting view's window-resize pass runs
-        // from windowDidLayout, so one layout alone can read as a false pass.
-        for _ in 0..<5 {
-            window.displayIfNeeded()
-            window.contentView?.layoutSubtreeIfNeeded()
-            RunLoop.current.run(until: Date().addingTimeInterval(0.02))
-        }
-
-        XCTAssertEqual(
-            window.frame.width, 500, accuracy: 1.0,
-            "Window width must stay where it was set — content ideal size must not grow the window"
-        )
-        XCTAssertEqual(
-            window.frame.height, 400, accuracy: 1.0,
-            "Window height must stay where it was set — content ideal size must not grow the window"
-        )
-    }
-
-    /// Same contract when the window sits BELOW the content's minimum size —
-    /// the live trigger: a programmatic resize can place a window under the
-    /// workspace chrome's minimum width, and the hosting view must not march
-    /// the window frame toward (or past) the content minimum in response.
-    @MainActor
-    func testWindowDoesNotGrowWhenSetBelowContentMinimumSize() {
-        let window = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 500, height: 400),
-            styleMask: [.titled, .closable, .resizable],
-            backing: .buffered,
-            defer: false
-        )
-        defer {
-            NotificationCenter.default.post(name: NSWindow.willCloseNotification, object: window)
-            window.orderOut(nil)
-        }
-        let wide = Color.clear.frame(
-            minWidth: 900, maxWidth: .infinity,
-            minHeight: 700, maxHeight: .infinity
-        )
-        window.contentView = MainWindowHostingView(rootView: AnyView(wide))
-        window.setFrame(NSRect(x: 0, y: 0, width: 500, height: 400), display: true)
-        window.makeKeyAndOrderFront(nil)
-
-        for _ in 0..<5 {
-            window.displayIfNeeded()
-            window.contentView?.layoutSubtreeIfNeeded()
-            RunLoop.current.run(until: Date().addingTimeInterval(0.02))
-        }
-
-        XCTAssertEqual(
-            window.frame.width, 500, accuracy: 1.0,
-            "Window width must stay where it was set even below the content minimum"
-        )
-        XCTAssertEqual(
-            window.frame.height, 400, accuracy: 1.0,
-            "Window height must stay where it was set even below the content minimum"
-        )
-    }
-}
-
 @MainActor
 final class TerminalWindowPortalLifecycleTests: XCTestCase {
     private final class ContentViewCountingWindow: NSWindow {
@@ -4793,7 +4703,7 @@ final class TerminalWindowPortalLifecycleTests: XCTestCase {
         }
     }
 
-    private func realizeWindowLayout(_ window: NSWindow) {
+    func realizeWindowLayout(_ window: NSWindow) {
         window.makeKeyAndOrderFront(nil)
         window.displayIfNeeded()
         window.contentView?.layoutSubtreeIfNeeded()
@@ -4801,7 +4711,7 @@ final class TerminalWindowPortalLifecycleTests: XCTestCase {
         window.contentView?.layoutSubtreeIfNeeded()
     }
 
-    private func drainMainQueue() {
+    func drainMainQueue() {
         let expectation = XCTestExpectation(description: "drain main queue")
         DispatchQueue.main.async {
             expectation.fulfill()
@@ -4884,64 +4794,6 @@ final class TerminalWindowPortalLifecycleTests: XCTestCase {
         terminalPortal.synchronizeHostedViewForAnchor(anchor)
 
         assertHostOrder("Terminal portal bind/sync should not rise above the browser portal host")
-    }
-
-    @MainActor
-    func testPortalSkipsSynchronousRefreshForHiddenSurfaces() throws {
-        let window = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 520, height: 340),
-            styleMask: [.titled, .closable],
-            backing: .buffered,
-            defer: false
-        )
-        defer {
-            NotificationCenter.default.post(name: NSWindow.willCloseNotification, object: window)
-            window.orderOut(nil)
-        }
-        realizeWindowLayout(window)
-        guard let contentView = window.contentView else {
-            XCTFail("Expected content view")
-            return
-        }
-
-        let portal = WindowTerminalPortal(window: window)
-        let visibleAnchor = NSView(frame: NSRect(x: 8, y: 8, width: 240, height: 160))
-        let hiddenAnchor = NSView(frame: NSRect(x: 260, y: 8, width: 240, height: 160))
-        contentView.addSubview(visibleAnchor)
-        contentView.addSubview(hiddenAnchor)
-
-        let visibleSurface = TerminalSurface(
-            tabId: UUID(), context: GHOSTTY_SURFACE_CONTEXT_SPLIT,
-            configTemplate: nil, workingDirectory: nil
-        )
-        let hiddenSurface = TerminalSurface(
-            tabId: UUID(), context: GHOSTTY_SURFACE_CONTEXT_SPLIT,
-            configTemplate: nil, workingDirectory: nil
-        )
-        portal.bind(hostedView: visibleSurface.hostedView, to: visibleAnchor, visibleInUI: true)
-        portal.bind(hostedView: hiddenSurface.hostedView, to: hiddenAnchor, visibleInUI: false)
-        portal.synchronizeHostedViewForAnchor(visibleAnchor)
-        drainMainQueue()
-        realizeWindowLayout(window)
-
-        visibleSurface.resetDebugForceRefreshCount()
-        hiddenSurface.resetDebugForceRefreshCount()
-
-        // Move BOTH anchors: both hosted views get geometry bookkeeping, but
-        // only the visible one may pay for the synchronous redraw — one
-        // layout pass syncs every hosted view in the window, and a mirror
-        // workspace parks 20+ surfaces on unselected tabs.
-        visibleAnchor.setFrameSize(NSSize(width: 220, height: 150))
-        hiddenAnchor.setFrameSize(NSSize(width: 220, height: 150))
-        portal.synchronizeHostedViewForAnchor(visibleAnchor)
-        drainMainQueue()
-
-        XCTAssertEqual(
-            hiddenSurface.debugForceRefreshCount(),
-            0,
-            "A hidden (unselected-tab) surface must not receive the synchronous GPU-blocking refresh on geometry sync"
-        )
-        withExtendedLifetime((visibleSurface, hiddenSurface)) {}
     }
 
     func testRegistryPrunesPortalWhenWindowCloses() {
